refactor: refactor inject functions in separate files
[AP-773]
